# perf: optimise `Modifiers`, remove linq and array allocations