- Use non boxing `ToArray` method.
- Use `Array.Sort` instead of `OrderBy` 
  - This could be a mistake. `OrderBy` is a stable sort whereas `Array.Sort` is unstable. If I understand correctly, this won't be a problem as csharpier will never encounter a situation where `SyntaxTokenList` contains 2 modifiers with the same name (but potentially different trivia). I don't know if preprocessor directives could break this 
- Replace `Zip().Any()` with a custom `SequenceEquals` (this prevents boxing `SyntaxTokenList`)
- Remove unnecessary `ToArray` call

Saves 1 MB on Tests benchmark and 1.8 MB on the Complex benchmark. This is a lot more that I expected, dotmemory claims shows a 200KB reduction but the benchmarks think otherwise 🤔 

#### Before
![image](https://github.com/user-attachments/assets/ac10377c-edb0-440d-957f-9e1372564457)

#### After
![image](https://github.com/user-attachments/assets/75165b03-1afa-4a7a-8b2a-d947e036ab9e)



### Benchmarks (on battery power timing is inaccurate)
#### Before
| Method                        | Mean     | Error   | StdDev   | Median   | Gen0      | Gen1      | Gen2      | Allocated |
|------------------------------ |---------:|--------:|---------:|---------:|----------:|----------:|----------:|----------:|
| Default_CodeFormatter_Tests   | 138.0 ms | 2.92 ms |  8.43 ms | 137.2 ms | 3000.0000 | 1000.0000 |         - |  32.02 MB |
| Default_CodeFormatter_Complex | 295.4 ms | 7.06 ms | 20.49 ms | 302.7 ms | 6000.0000 | 3000.0000 | 1000.0000 |  56.42 MB |


#### After
| Method                        | Mean     | Error   | StdDev   | Gen0      | Gen1      | Allocated |
|------------------------------ |---------:|--------:|---------:|----------:|----------:|----------:|
| Default_CodeFormatter_Tests   | 137.8 ms | 2.61 ms |  5.89 ms | 2000.0000 | 1000.0000 |  31.07 MB |
| Default_CodeFormatter_Complex | 286.0 ms | 5.42 ms | 10.31 ms | 5000.0000 | 2000.0000 |  54.61 MB |